Gates user

+++ ecommerce/app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -64,6 +65,11 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if (Gate::denies('edit-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
+
         $roles = Role::all();
 
         return view('admin.users.edit', [
@@ -93,6 +99,11 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if (Gate::denies('delete-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
+
         $user->roles()->detach();
         $user->delete();
